Add feature to connect to master server
A new feature has been added that allows a slave replica to connect to the master server. In the app/main.py file, a function has been defined to initiate this connection. If the server is marked as a 'slave', it will attempt communication with the master server using socket connection.