Implement optional replication configuration
An optional replication configuration has been put in place in the main.py file. This includes the inclusion of 'slave hosts' parameter and the enforcement of responses from the master socket with the REPLCONF command. Functions are also added to handle replica commands acting as a framework for handling and responding to different replication configuration commands from slave hosts.